This is crazy
I suddenly started getting linker errors when I compiled the objects
with -O3. Turns out I used a templated function that had the definition
in a C++ file. Good job.

+++ src/lexer.hpp
@@ -14,7 +14,7 @@ private:
   void skip_whitespace();
   std::string read_identifier();
   std::string input;
-  int position{0};
-  int read_position{0};
+  size_t position{0};
+  size_t read_position{0};
   char ch{0};
 };
